QmlDesigner: Add simple cache for font family names
Reading a single font family name can take up to 5ms and for
multiple fonts this slows down the property editor significantly.
Since the font families for a specific font do not change, we simply
cache them.

+++ src/plugins/qmldesigner/components/propertyeditor/fontresourcesmodel.cpp
@@ -57,6 +57,13 @@ const QStringList &fontFilesFilterList()
 
 std::optional<QString> fontFamily(const QString &fontPath)
 {
+    static QHash<QString, QString> s_cache;
+
+    auto cached_name = s_cache.value(fontPath);
+
+    if (!cached_name.isEmpty())
+        return cached_name;
+
     const int fontId = QFontDatabase::addApplicationFont(fontPath);
     const QStringList fontFamilies = QFontDatabase::applicationFontFamilies(fontId);
     QFontDatabase::removeApplicationFont(fontId);
@@ -68,6 +75,8 @@ std::optional<QString> fontFamily(const QString &fontPath)
         ProjectExplorer::TaskHub::addTask(issueType, issueDesc, issueCategory);
         return {};
     }
+
+    s_cache.insertOrAssign(fontPath, fontFamilies.front());
 
     return fontFamilies.front();
 }
